Multiple Leg floor detection implemented
The player now fires multiple rays from the base of their body to detect the floor. This is used to prevent a single hole from causing the player to fall and to have a more smooth transition when going up straight stairs. It also will only cause a player to start falling if a certain amount of "legs" do not connect to an object.